refactor Ant into ant.rs
