+++ src/aco/ants_system/builder.rs
@@ -1,6 +1,7 @@
 use crate::aco;
+use crate::aco::ant::Ant;
 use crate::aco::ant_system_cfg::AntSystemCfgOpt;
-use crate::aco::ants_system::{Ant, Solution};
+use crate::aco::ants_system::Solution;
 use crate::aco::pheromone::PheromoneUpdate;
 use crate::aco::probe::Probe;
 use crate::aco::{AntSystem, AntSystemCfg, FMatrix};
